+++ app/src/main/java/com/deutschdreamers/wordimpostor/ui/navigation/Screen.kt
@@ -32,5 +32,8 @@ sealed interface Screen {
 
     @Serializable
     data object About : Screen
+
+    @Serializable
+    data object Stats : Screen
 }